requirements.txt
yt-dlp>=2023.12.30
spotipy>=2.23.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
argparse
